requirements.txt
beautifulsoup4==4.4.1
diff-match-patch==20121119
Django==1.9.6
django-allauth==0.25.2
django-appconf==1.0.2
django-ckeditor==5.0.3
django-classy-tags==0.7.2
django-crispy-forms==1.6.0
django-formtools==1.0
django-mptt==0.8.4
django-reversion==1.10.2
Django-Select2==5.8.5
django-taggit==0.19.1
djangorestframework==3.3.3
funcsigs==0.4
Markdown==2.6.6
matplotlib==1.4.3
mock==1.3.0
ndg-httpsclient==0.4.0
nose==1.3.7
numpy==1.10.1
oauthlib==1.1.1
pbr==1.7.0
pyasn1==0.1.8
PyBindGen==0.17.0.post41+ngd10fa60
pygccxml==1.6.2
pyparsing==2.0.3
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.4
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
virtualenv==13.1.0
